+++ cmd/get/task.go
@@ -2,6 +2,7 @@ package get
 
 import (
 	"context"
+	"encoding/json"
 	"github.com/lyft/flytestdlib/logger"
 
 	"github.com/lyft/flytectl/pkg/adminutils"
@@ -12,6 +13,23 @@ import (
 
 	"github.com/lyft/flyteidl/gen/pb-go/flyteidl/admin"
 )
+
+var taskStructure = map[string]string{
+	"Version":          "$.id.version",
+	"Name":             "$.id.name",
+	"Type":             "$.closure.compiledTask.template.type",
+	"Discoverable":     "$.closure.compiledTask.template.metadata.discoverable",
+	"DiscoveryVersion": "$.closure.compiledTask.template.metadata.discovery_version",
+}
+
+var transformTask = func(jsonbody []byte) (interface{}, error) {
+	results := PrintableTask{}
+	if err := json.Unmarshal(jsonbody, &results); err != nil {
+		return results, err
+	}
+	return results, nil
+}
+
 
 func getTaskFunc(ctx context.Context, args []string, cmdCtx cmdCore.CommandContext) error {
