+++ cmd/version.go
@@ -3,7 +3,6 @@ package cmd
 import (
 	"github.com/lyft/flytestdlib/version"
 	"github.com/spf13/cobra"
-	"github.com/lyft/flytestdlib/blob/master/version"
 )
 
 var (
